+++ apps/meteor/app/nextcloud/client/index.ts
@@ -1,1 +1,1 @@
-import '../lib/common';
+import './lib';
