+++ apps/meteor/app/nextcloud/client/lib.ts
@@ -1,0 +1,40 @@
+import { Meteor } from 'meteor/meteor';
+import { Tracker } from 'meteor/tracker';
+import _ from 'underscore';
+import type { OauthConfig } from '@rocket.chat/core-typings';
+
+import { settings } from '../../settings/client';
+import { CustomOAuth } from '../../custom-oauth/client/custom_oauth_client';
+
+const config: OauthConfig = {
+	serverURL: '',
+	tokenPath: '/index.php/apps/oauth2/api/v1/token',
+	tokenSentVia: 'header',
+	authorizePath: '/index.php/apps/oauth2/authorize',
+	identityPath: '/ocs/v2.php/cloud/user?format=json',
+	scope: 'openid',
+	addAutopublishFields: {
+		forLoggedInUser: ['services.nextcloud'],
+		forOtherUsers: ['services.nextcloud.name'],
+	},
+};
+
+const Nextcloud = new CustomOAuth('nextcloud', config);
+
+const fillServerURL = _.debounce((): void => {
+	const nextcloudURL = settings.get('Accounts_OAuth_Nextcloud_URL');
+	if (!nextcloudURL) {
+		if (nextcloudURL === undefined) {
+			return fillServerURL();
+		}
+		return;
+	}
+	config.serverURL = nextcloudURL.trim().replace(/\/*$/, '');
+	return Nextcloud.configure(config);
+}, 100);
+
+Meteor.startup(function () {
+	Tracker.autorun(function () {
+		return fillServerURL();
+	});
+});
